Re-Added EnableS3Trick()
This one was missing in between cherry picking changed from local branch. It has been re-added.

In addition the default value of this option, if not in the config file is "false. This is because as per MS KB this only applies to legacy systems and only under XP. On kernel 6.1 system this registry key as no effect.