Fixed Mouse Hiding issues. Updated OnResume()/OnSuspend()
MP will now minimize itself to the tray on a suspend request and on a subsequent resume maximizes itself again. Needs some testing in detail.

In addition an imbalance between Cursor.Hide()/Cursor.Show() was spotted.

If you are in fullscreen and the mouse cursor was never showed and then minimize MP to the tray, the mouse cursor would hide as soon as you open the context menu.

This was solved by using Win32 API's ShowCursor in MinimizeToTray() and RestoreFromTray().

The full mouse cursor handling needs to be updated to D3D. The current mix of Windows Forms and D3D in Fullscreen Mode (including the splash screen) has too many side-effects and uncontrolled states requiring workarounds.