Fixed startup problems in fullscreen.
When we create a D3D full-screen device in a border-less full-screen WinForm the screen goes to black, because we haven't rendered anything yet with D3D. There is no way to circumvent this with when we use pure C# for the device creation.

Hiding the form, so that you can see the Splash-Screen (running in a different form, in another thread) cannot be done by the usual form controls at this point, because the full form handling is not yet in effect at this point it time. In other words: we are not active until some resources are loaded.

The current master first creates a windowed D3D device as a workaround and later switches to full-screen. This results in longer startup times as the resources are loaded two times. Once just for the sake of splash-screen and then for real.

To solve that I use a call to ShowWindow(Handle, SW_MINIMIZE) for the Win32 API to minimize the window. Hey, this works great. The empty D3D screen goes into the background and the splash screen is visible.

And now the kicker: This doesn't work if NO OTHER window on the Desktop is visible.

As soon as there is at least one window on the Desktop it works fine and the ShowWindow() call doesn't hang.

Now why does it? It seem to be caused by the Splash-Screen Form itself. I am not big on debugging this one as I intend to replace it with proper D3D calls anyway. But there must be something inside its form handling, probably because the thread doesn't do any message handling.

Solution for now: Move the SplashScreen.Run() call after creating the DirectX device and everything works fine