Mantis #3728
Added a check for the existence of MediaPortal.xml in CreateTemporaryConfiguration. When not found Watchdog.exe will start Configuration.exe in Wizard mode, wait for configuration to be completed and restarts the Watchdog App.

In addition potential null reference check has been added to CheckRequirements.

Removal of a few redundant statements.