Mouse Cursor will only be visible again when the mouse really moved and not only based on events.
Minor enhancements to when save/resume the player state and when to pause/resume playing.

When ever a new form is generated or initialised, Windows send an OnMouseMove Event to the form.

That causes the appearance of the mouse pointer in various circumstances (e.g. starting of the DVD player).

To avoid this. The OnMouseMoveEvent Method now tracks the cursor position of the last event and only makes the mouse cursor visible again when the position of the current and the last event are different.

In addition the locations where the player state is saved/resumed and when playback is paused/resumed where updated.

Current problem: cannot pause DVD menu playback yet.