Removed all D3D device resets
They are causing too much issues and are not needed unless the device is really lost.

Also removed the toggling of DirectX Exclusive mode for VMR9. It basically tries to setup a real D3D fullscreen instead of a borderless window. Applies to XP only. Will be re-added later when dealing with XP compability once no issues are reported anymore.