Mantis #3808
textColorNoFocus, textColorNoFocus and textColorNoFocus2 were added to GUIListControl.cs

If not defined they take the value of their textcolor counterpart.

This property will override the textcolor for items that are not selected.

In addition the method RenderLabel() was commented a received a bit of a cleanup. It's still very ugly but not at lest a bit understandable for future changes.