Mantis #4193
Allow out of bounds rendering for a label when the label itself don't defined bounds.

This is for backwards compatibility with labels not having set width/height but still want to do a center alignment relatively to their text width. Meaning shifting it half the text length to the left for example (applies to watched count/watched percentage in current skins)