Changing RenderButton() of GUIListControl.cs to make use of unfocusedAlpha again.
We might deal here with an unfinished implementation of alpha blending textures as the RederButton() function uses a variable _diffuseColor of its parent class.

This variable is changed by the Opacity() method of the parent class. But never gets called anywhere in the code base.

It's possible that other classes are also affected. This is an initial fix for testing if RenderButton() is the actual problem.