Updating Clipping in GUIFadeLabel to prevent wrong clipping when a negative vertical offset is used.
In some cases it can happen that a clipping regions position does not match the position of the text anymore.

This happens during ALIGN_CENTER of GUIFadeLabel.cs when no height is given. Therefore not the initial vertical position need sot be taken but the updated one as they now differ.

There is still one glitch left, where a scrolling line can jump up a few pixels once the scrolling beings. This is caused by the fact, that GUIFadeLabel, GUILabelControl, and GUIControl all manipulate the vertical position differently.

It is only visible when the vertical height is smaller than the one of the skin itself.  Happens because of constant type casting between int, float and double instead of proper use of rounding functions or using a dynamic variable and let C# automatically handle this properly.