LIVE555: remove modification 1 and associated code.
The original reason for this code was to prevent paused [RTSP] streams - both live and recorded/static??? - on a client from being kicked by the streaming server. Kicking is triggered after 65 seconds in which the server hasn't received any RTSP communication (any command will do) or RTCP receiver reports from the client.

In theory, TsReader sends an RTSP DESCRIBE command to the server every ~5 seconds in order to get duration updates. That applies for live TV, live radio and in-progress recordings, and does not stop even when streaming is paused. There's no such communication when streaming completed recordings. However, testing shows that paused streams for static recordings aren't dropped [after 65 seconds] either. This appears to be because LIVE555 sends regular RTCP receiver reports, which also don't stop even when streaming is paused.

In short: it appears that this code is obsolete and can safely be removed.

One last LIVE555 modification in MPEG1or2Demux.cpp remains.